--- definitions/cardtypes.js
@@ -1,34 +1,0 @@
-module.exports = {
-  guard: {
-    quantity: 5,
-    fields: ["target", "guess"]
-  },
-  priest: {
-    quantity: 2,
-    fields: ["target"]    
-  },
-  baron: {
-    quantity: 2,
-    fields: ["target"] 
-  },
-  handmaid: {
-    quantity: 2,
-    fields: []
-  },
-  prince: {
-    quantity: 2,
-    fields: ["target"]
-  },
-  king: {
-    quantity: 1,
-    fields: ["target"]
-  },
-  countess: {
-    quantity: 1,
-    fields: []
-  },
-  princess: {
-    quantity: 1,
-    fields: []
-  }
-};
