+++ model/cards/princess.js
@@ -1,11 +1,18 @@
 class Princess {
-  constructor() {    
+  constructor() {
     this.value = 8;
     this.name = "princess";
   }
 
   play(player) {
     player.kill();
+  }
+
+  info() {
+    return {
+      name: this.name,
+      value: this.value
+    };
   }
 }
