+++ package.json
@@ -10,10 +10,9 @@
   "license": "ISC",
   "dependencies": {
     "bluebird": "^3.5.0",
-    "body-parser": "^1.17.1",
     "card-deck": "^1.0.0",
-    "express": "^4.15.2",
     "lodash": "^4.17.4",
+    "lovebotplayer": "^1.0.1",
     "request": "^2.80.0",
     "request-promise": "^4.1.1"
   }
