--- playerapi/playerapi.js
@@ -1,40 +1,0 @@
-let express = require('express')
-let bodyParser = require('body-parser')
-let _ = require('lodash')
-let random = require('../strategy/random');
-let goodguess = require('../strategy/goodguess');
-
-
-module.exports = {
-    start: (logRequests) => {
-        let strategy = goodguess;
-
-        let app = express();
-
-        app.use(bodyParser.json())
-
-        if (logRequests) {
-            app.use(function(req, res, next) {
-                console.log(
-                    "API for '" + 
-                    req.body.player.number + 
-                    "' received call: " + 
-                    JSON.stringify(req.body.player).substring(0,60) + "...");
-                next();
-            });
-        }
-
-        app.post('/api/takeTurn', (req, res) => {
-            strategy(req.body.player, req.body.opponents)
-                .then(result => res.json(result));
-        })
-
-        let port = 8080;
-        app.listen(port, function() {
-            console.log('Player api listening on port: ' + port)
-        });
-    }
-}
-
-
-
